requirements.txt
python-telegram-bot==21.*
aiomysql==0.2.*
python-dotenv==1.*
